implement book listing
The reseller registers all consumers in its book listing. For simplicity
this happens once in the beginning and no consumers can be added or
removed later on.

When the reseller gets new book licenses they can enlist the new book on
their listing. When a book is sold out they can unlist it.